// .pnpm-store/v11/projects/1d17c93f926bb33b78ccc0867f7dd813/src/types/api.types.ts
export interface Student {
  id: string;
  admissionNumber: string;
  firstName: string;
  lastName: string;
  dateOfBirth?: string;
  gender?: string;
  passportPhoto?: string;
  isActive: boolean;
  academicYearId?: string;
  termId?: string;
  classId?: string;
  studentCategoryId?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Payment {
  id: string;
  studentId: string;
  amount: number;
  method: string;
  status: string;
  description?: string | null;
  date: string;
  student?: Student;
  createdAt: string;
  updatedAt: string;
}